moved SpectralProfilePlotXAxisUnitModel to  spectrallibraryplotunitmodels.py
